Add core module exports, shared constants, and act templates
- Implement core module exports in index.ts
- Create shared constants for application configuration
- Develop shared utilities including spinner management
- Define types for various models and provider configurations
- Introduce act templates for diverse roles in the template module
- Implement utility functions for listing and applying act templates
