+++ src/cli/commands/provider.command.ts
@@ -2,7 +2,10 @@ import {select} from '@inquirer/prompts';
 import chalk from 'chalk';
 import Table from 'cli-table3';
 import {Command} from 'commander';
-import {getAllProviders, getProvider} from '../../config/providers/index.js';
+import {
+  getAllProviders,
+  getProvider,
+} from '../../core/config/providers/index.js';
 
 export const createProviderCommand = (): Command => {
   const command = new Command('provider');
