+++ src/core/config/manager.ts
