+++ src/core/config/providers/amazon-bedrock.provider.ts
