+++ src/core/config/providers/ollama.provider.ts
