+++ src/core/config/providers/openai.provider.ts
