+++ src/core/config/providers/qwen.provider.ts
