+++ src/core/database/adapter.ts
