+++ src/core/database/config/paths.ts
