+++ src/core/index.ts
@@ -1,0 +1,6 @@
+// Core module exports
+
+export * from './config/index.js';
+export * from './database/index.js';
+export * from './shared/index.js';
+export * from './template/index.js';
