+++ src/core/shared/constants.ts
