+++ src/core/shared/index.ts
