+++ src/core/template/index.ts
