+++ src/core/template/utils.ts
